requirements.txt
aiogram~=2.14.3
APScheduler~=3.7.0
python-dotenv~=0.19.0
pytz~=2021.1
gspread~=5.3.0
SQLAlchemy~=1.4.36
PyJWT~=2.4.0
requests~=2.27.1
sqlalchemy~=1.4.36